restore env example and clean doc (#17)
