--- src/frontend/src/clerk/clerk-logout.ts
@@ -1,29 +1,0 @@
-import { IS_CLERK_AUTH } from "@/constants/clerk";
-import { useLogout as useLogoutMutation } from "@/controllers/API/queries/auth";
-import { useClerk } from "@clerk/clerk-react";
-
-export function useLogout(options?: Parameters<typeof useLogoutMutation>[0]) {
-  const { mutate, mutateAsync, ...rest } = useLogoutMutation(options);
-  const { signOut } = IS_CLERK_AUTH ? useClerk() : { signOut: async () => {} };
-
-  const clerkSignOut = async () => {
-    if (IS_CLERK_AUTH) {
-      try {
-        await signOut();
-      } catch (err) {
-        console.error("Clerk signOut failed:", err);
-      }
-    }
-  };
-
-  const wrappedMutate: typeof mutate = (...args) => {
-    clerkSignOut().finally(() => mutate(...args));
-  };
-
-  const wrappedMutateAsync: typeof mutateAsync = async (...args) => {
-    await clerkSignOut();
-    return mutateAsync(...args);
-  };
-
-  return { mutate: wrappedMutate, mutateAsync: wrappedMutateAsync, clerkSignOut, ...rest };
-}
