+++ src/frontend/src/controllers/API/queries/auth/use-post-refresh-access.ts
@@ -1,5 +1,4 @@
-import { IS_CLERK_AUTH } from "@/constants/clerk";
-import { mockClerkMutation } from "@/utils/mocks/mockMutationClerk";
+import { IS_CLERK_AUTH, mockClerkMutation } from "@/clerk/auth";
 import { IS_AUTO_LOGIN, LANGFLOW_REFRESH_TOKEN } from "@/constants/constants";
 import useAuthStore from "@/stores/authStore";
 import { useMutationFunctionType } from "@/types/api";
