--- src/frontend/src/pages/ClerkLoginPage/index.tsx
@@ -1,4 +1,0 @@
-import { SignIn } from "@clerk/clerk-react";
-export default function ClerkLoginPage() {
-  return <SignIn />;
-}
