+++ src/frontend/src/routes.tsx
@@ -33,7 +33,7 @@ import GlobalVariablesPage from "./pages/SettingsPage/pages/GlobalVariablesPage"
 import MessagesPage from "./pages/SettingsPage/pages/messagesPage";
 import ShortcutsPage from "./pages/SettingsPage/pages/ShortcutsPage";
 import ViewPage from "./pages/ViewPage";
-import { LoginPage, SignUp, LoginAdminPage } from "./routes/login-routes";
+import { LoginPage, SignUp, LoginAdminPage } from "./clerk/login-pages";
 
 const AdminPage = lazy(() => import("./pages/AdminPage"));
 const DeleteAccountPage = lazy(() => import("./pages/DeleteAccountPage"));
