--- src/frontend/src/utils/mocks/mockMutationClerk.ts
@@ -1,13 +1,0 @@
-export const mockClerkMutation = {
-  mutate: () => {},
-  mutateAsync: async () => undefined,
-  isError: false,
-  isIdle: true,
-  isPending: false,
-  isSuccess: true,
-  reset: () => {},
-  status: "success",
-  variables: undefined,
-  data: undefined,
-  error: null,
-} as any;
